+++ app/src/main/java/db/ecommerce/model/Shopping.java
@@ -10,6 +10,6 @@ public interface Shopping {
     /**
      * @return the price
      */
-    float getPrice();
+    double getPrice();
 
 }
